+++ integration-tests/zombie-tests/src/main.spec.ts
@@ -14,6 +14,7 @@ import { start } from "polkadot-api/smoldot"
 import {
   AccountId,
   Binary,
+  CompatibilityLevel,
   PolkadotClient,
   SS58String,
   TxEvent,
@@ -22,7 +23,7 @@ import {
 import { getSmProvider } from "polkadot-api/sm-provider"
 import { getWsProvider } from "polkadot-api/ws-provider/node"
 import { createClient as createRawClient } from "@polkadot-api/substrate-client"
-import { MultiAddress, roc } from "@polkadot-api/descriptors"
+import { getMetadata, MultiAddress, roc } from "@polkadot-api/descriptors"
 import { accounts } from "./keyring"
 import { getPolkadotSigner } from "polkadot-api/signer"
 import { fromHex } from "@polkadot-api/utils"
@@ -67,6 +68,7 @@ describe("E2E", async () => {
   } else {
     client = createClient(
       withPolkadotSdkCompat(getWsProvider("ws://127.0.0.1:9934")),
+      { getMetadata },
     )
   }
   console.log("client started")
@@ -306,6 +308,42 @@ describe("E2E", async () => {
     )
   })
 
+  it("different mortality values", async () => {
+    const alice = accounts["alice"]["sr25519"]
+    const bob = accounts["bob"]["sr25519"]
+    const aliceAddress = accountIdDec(alice.publicKey)
+    const bobAddress = accountIdDec(bob.publicKey)
+    const initialNonce =
+      await api.apis.AccountNonceApi.account_nonce(aliceAddress)
+
+    const tx = api.tx.Balances.transfer_allow_death({
+      dest: MultiAddress.Id(bobAddress),
+      value: ED,
+    })
+    let i = 0
+    const submitAndWait = (
+      mortality: { mortal: false } | { mortal: true; period: number },
+    ) =>
+      lastValueFrom(
+        tx.signSubmitAndWatch(alice, { mortality, nonce: initialNonce + i++ }),
+      )
+
+    await Promise.all([
+      submitAndWait({ mortal: false }),
+      submitAndWait({ mortal: true, period: 8 }),
+      submitAndWait({ mortal: true, period: 25 }),
+      submitAndWait({
+        mortal: true,
+        period: api.constants.System.BlockHashCount(token),
+      }),
+    ])
+
+    const currentNonce =
+      await api.apis.AccountNonceApi.account_nonce(aliceAddress)
+
+    expect(currentNonce).toEqual(initialNonce + 4)
+  })
+
   it("keeps on validating transactions after they have been broadcasted", async () => {
     const alice = accounts["alice"]["sr25519"]
     const bob = accounts["bob"]["sr25519"]
@@ -420,9 +458,11 @@ describe("E2E", async () => {
   it("queries opaque storage entries", async () => {
     // some old polkadot-sdk versions don't include this pallet
     // ensure that some version tested include it
-    try {
-      await api.query.CoretimeAssignmentProvider.CoreDescriptors.getCompatibilityLevel()
-    } catch {
+    if (
+      !(await api.query.CoretimeAssignmentProvider.CoreDescriptors.isCompatible(
+        CompatibilityLevel.Partial,
+      ))
+    ) {
       return
     }