+++ packages/client/src/tx/signed-extensions/mortal-enc.ts
@@ -9,12 +9,16 @@ function trailingZeroes(n: number) {
   return i
 }
 
+const nextPower = (n: number) => 1 << Math.ceil(Math.log2(n))
+
 export const mortal = enhanceEncoder(
   Bytes(2)[0],
-  (value: { period: number; phase: number }) => {
-    const factor = Math.max(value.period >> 12, 1)
-    const left = Math.min(Math.max(trailingZeroes(value.period) - 1, 1), 15)
-    const right = (value.phase / factor) << 4
+  (value: { period: number; startAtBlock: number }) => {
+    const period = Math.min(Math.max(nextPower(value.period), 4), 1 << 16)
+    const phase = value.startAtBlock % period
+    const factor = Math.max(period >> 12, 1)
+    const left = Math.min(Math.max(trailingZeroes(period) - 1, 1), 15)
+    const right = (phase / factor) << 4
     return u16[0](left | right)
   },
 )