+++ packages/merkleize-metadata/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 1.1.17 - 2025-05-20
+
+### Fixed
+
+- Update dependencies
+
 ## 1.1.16 - 2025-05-15
 
 ### Fixed