+++ packages/smoldot/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.3.9 - 2025-05-27
+
+### Fixed
+
+- Update smoldot to `2.0.35`
+
 ## 0.3.8 - 2024-12-10
 
 ### Fixed